moved tests and edited scripts to work with packaged module
