+++ build_docker
@@ -8,13 +8,14 @@ mkdir -p $tmpDir
 rm -rf $tmpDir/*
 
 cp Dockerfile $tmpDir/
-cp ShapeShifter/RunTests.sh $tmpDir/
-cp ShapeShifter/RunTests2.sh $tmpDir/
+cp Tests/RunTests.sh $tmpDir/
+cp Tests/RunTests2.sh $tmpDir/
 cp VERSION $tmpDir/
 mkdir $tmpDir/Tests/
-cp -r ShapeShifter/Tests/InputData $tmpDir/Tests/
-cp -r ShapeShifter/Tests/OutputData $tmpDir/Tests/
-cp ShapeShifter/*.py $tmpDir
+cp -r Tests/InputData $tmpDir/Tests/
+cp -r Tests/OutputData $tmpDir/Tests/
+cp Tests/*.py $tmpDir
+cp -r shapeshifter $tmpDir
 cd $tmpDir
 docker build -t srp33/shapeshifter:version$(cat VERSION) .
