+++ shapeshifter/utils/__init__.py
@@ -1,7 +1,7 @@
 __all__ = ['standardize_null_value', 'is_gzipped', 'build_parquet', 'salmon_to_pandas', 'kallisto_to_pandas',
            'to_arff', 'to_gct', 'gct_to_pandas', 'is_date', 'is_numeric', 'arff_to_pandas']
 
-from shapeshifter.utils.comparedataframes import standardize_null_value
+from Tests.CompareDataframes import standardize_null_value
 from shapeshifter.utils.convertarff import is_date, is_numeric, arff_to_pandas, to_arff
 from shapeshifter.utils.convertgct import to_gct, gct_to_pandas
 from shapeshifter.utils.kallisto import kallisto_to_pandas
